Add difficulty settings.
The difficulty only affects the scoring and the penalty for a missed
scan.  For `Rusty`, you get half points and the penalty is an unscored
hard drop.  For `Trusty`, you get standard points and the penalty is
garbage lines.  For `Lusty`, you get double points and the penalty is a
game over.